requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.12
langchain-openai==0.1.5
langchain-google-genai==0.0.9
langchain-community==0.0.24
langchain-chroma==0.1.1
langchain-text-splitters==0.0.1
pydantic==2.6.3
python-dotenv==1.0.1
chromadb==0.4.22
openai==1.12.0
google-generativeai==0.3.2